Move credentialStatus and credentialSchema to their own sections.
